fixed topic and slowmode bugs

+++ src/commands/mod/topic.js
@@ -2,8 +2,14 @@ const PREFIX = process.env.PREFIX;
 const fs = require('fs');
 
 module.exports = {
-    run: async (client, message, args) => {
-        message.delete()
+    run: async (client, message, args) => { 
+        if (message.guild.me.hasPermission('MANAGE_MESSAGES')) message.delete();
+        if (!message.guild.me.hasPermission('MANAGE_CHANNELS')) {
+            return message.channel.send(":x: **I need the `Manage Channels` permission change the topic of this channel.**")
+            .then(msg => {
+                msg.delete({ timeout: 4000 });
+            });
+        }
         if (!message.member.hasPermission('MANAGE_CHANNELS')) {
             return message.channel.send(":x: **You need the `Manage Channels` permission change the topic of this channel.**")
             .then(msg => {
@@ -26,9 +32,10 @@ module.exports = {
                 if (msg.includes(bannedPhrases[j])) return;
             }
         }
-
         if (args == `${PREFIX}topic`) return;
-        message.channel.setTopic(args);
+        message.channel.setTopic(args)
+        .then(updated => message.channel.send(`<#${message.channel.id}>'s new topic is ${updated.topic}`))
+        .catch(console.error);
     },
     aliases: [],
     description: 'Sets the topic for a channel'
